+++ demo/src/my-app.ts
@@ -1,10 +1,10 @@
 // @ts-ignore
-import { ObjectDom, Div, H1, Button, Row } from "object-dom";
+import { ObjectDom, Div, H1, Button, Row, Canvas } from "object-dom";
 
 export class MyApp extends ObjectDom {
   render = () => {
     return new Div({
-      children: [new H1({ text: "Counter Example" }), new Counter()],
+      children: [new H1({ text: "Counter Example" }), new Counter(), new CanvasExample()],
     });
   };
 }
@@ -45,23 +45,23 @@ class Counter extends ObjectDom {
   }
 }
 
-// class CanvasExample extends ObjectDom {
-//   render() {
-//     return new Canvas({
-//       style: { width: "200px", height: "200px" },
-//       onCreate: (node) => {
-//         const canvas = node as HTMLCanvasElement;
-//         const ctx = canvas.getContext("2d")!;
+class CanvasExample extends ObjectDom {
+  render() {
+    return new Canvas({
+      style: { width: "200px", height: "200px" },
+      onCreate: (node) => {
+        const canvas = node as HTMLCanvasElement;
+        const ctx = canvas.getContext("2d")!;
 
-//         // Create gradient
-//         const grd = ctx.createRadialGradient(75, 50, 5, 90, 60, 100);
-//         grd.addColorStop(0, "red");
-//         grd.addColorStop(1, "white");
+        // Create gradient
+        var grd = ctx.createRadialGradient(75, 50, 5, 90, 60, 100);
+        grd.addColorStop(0, "red");
+        grd.addColorStop(1, "white");
 
-//         // Fill with gradient
-//         ctx.fillStyle = grd;
-//         ctx.fillRect(10, 10, 150, 80);
-//       },
-//     });
-//   }
-// }
+        // Fill with gradient
+        ctx.fillStyle = grd;
+        ctx.fillRect(10, 10, 150, 80);
+      },
+    });
+  }
+}