+++ src/transformers/render.ts
@@ -6,9 +6,9 @@ import type { ObjectDom } from "../base";
 export function render(source: ObjectDom<HTMLElement>, target: HTMLElement = document.body) {
   let node = source.render().node;
   source.update = () => {
-    if (node) node.remove();
-    node = source.render().node;
-    target.appendChild(node);
+    const _newNode = source.render().node;
+    target.parentNode?.replaceChild(_newNode, node);
+    node = _newNode;
   };
   target.appendChild(node);
 }